+++ lib/base/base.dart
@@ -11,84 +11,145 @@ import 'package:flutter/material.dart';
 import '../types.dart';
 import 'controller.dart';
 
-/// 用于创建一个通用的轮播组件，可以基于它封装各种轮播
+/// Used to create a generic carousel component that can be customized for various carousels.
 ///
+/// 用于创建通用的轮播组件，可为各种轮播自定义
+///
+/// - [units] represents a list of components to be carousel-ed.
 /// - [units] 表示要轮播的组件列表
+///
+/// - [height] represents the height of the carousel component, defaults to 400.0.
 /// - [height] 表示轮播组件的高度，默认为400.0
+///
+/// - [width] represents the width of the carousel component, defaults to 0.0.
 /// - [width] 表示轮播组件的宽度，默认为0.0
+///
+/// - [onUnitTapped] represents a callback function when a carousel unit is tapped, optional.
 /// - [onUnitTapped] 表示当轮播单元被点击时的回调函数，可选
+///
+/// - [onIndicatorTapped] represents a callback function when an indicator is tapped, optional.
 /// - [onIndicatorTapped] 表示当指示器被点击时的回调函数，可选
+///
+/// - [useindicator] indicates whether to use indicators, defaults to true.
 /// - [useindicator] 表示是否使用指示器，默认为true
+///
+/// - [indicatorColor] represents the default indicator color, defaults to white.
 /// - [indicatorColor] 表示默认指示器的颜色，默认为白色
+///
+/// - [currentIndicatorColor] represents the color of the currently selected indicator, defaults to blue.
 /// - [currentIndicatorColor] 表示当前选中指示器的颜色，默认为蓝色
+///
+/// - [indicatorWidth] represents the width of the indicators, defaults to 40.0.
 /// - [indicatorWidth] 表示指示器的宽度，默认为40.0
+///
+/// - [indicatorHeight] represents the height of the indicators, defaults to 26.0.
 /// - [indicatorHeight] 表示指示器的高度，默认为26.0
+///
+/// - [indicatorMargin] represents the spacing between the indicators, defaults to 3.0.
 /// - [indicatorMargin] 表示指示器之间的间距，默认为3.0
+///
+/// - [indicatorToBottom] represents the distance of the indicators from the bottom, defaults to 10.0.
 /// - [indicatorToBottom] 表示指示器距离底部的距离，默认为10.0
+///
+/// - [indicatorShape] represents the default shape of the indicators, defaults to a rectangle.
 /// - [indicatorShape] 表示默认指示器的形状，默认为矩形
+///
+/// - [pageSnapping] indicates whether to enable page snapping, defaults to true.
 /// - [pageSnapping] 表示是否启用页面快速吸附，默认为true
+///
+/// - [padEnds] indicates whether to add extra pages at the beginning and end of the carousel, default is true.
 /// - [padEnds] 表示是否在轮播首尾添加额外的页面，默认为true
+///
+/// - [controller] represents the carousel controller, optional.
 /// - [controller] 表示轮播控制器，可选
+///
+/// - [disableIndicatorDefaultCallbacks] Whether to disable default indicator callback functions.
+/// - [disableIndicatorDefaultCallbacks] 是否禁用默认的指示器回调函数
+///
 class Carousel extends StatefulWidget {
+  /// List of components to be carousel-ed.
   /// 要轮播的组件列表
   final List<Widget> units;
 
+  /// Height of the carousel component.
   /// 轮播组件的高度
   final double height;
 
+  /// Width of the carousel component.
   /// 轮播组件的宽度
   final double width;
 
-  /// 是否使用指示器，默认为 true
+  /// Whether to use indicators, defaults to true.
+  /// 是否使用指示器，默认为true
   final bool useindicator;
 
-  /// 当图像被点击时的回调函数，可选
+  /// Callback function when a unit in the carousel is clicked, optional.
+  /// 当轮播单元被点击时的回调函数，可选
   final FunctionWithAInt? onUnitTapped;
 
+  /// Callback function when an indicator is clicked, optional.
   /// 当指示器被点击时的回调函数，可选
   final FunctionWithAInt? onIndicatorTapped;
 
+  /// Custom indicator builder function, optional.
   /// 自定义指示器构建函数，可选
   final FunctionIndicatorBuilder? indicatorBuilder;
 
-  /// 默认指示器的颜色
+  /// Default indicator color.
+  /// 默认指示器颜色
   final Color indicatorColor;
 
+  /// Color of the currently selected indicator.
   /// 当前选中指示器的颜色
   final Color currentIndicatorColor;
 
-  /// 指示器的宽度
+  /// Indicator width.
+  /// 指示器宽度
   final double indicatorWidth;
 
-  /// 指示器的高度
+  /// Indicator height.
+  /// 指示器高度
   final double indicatorHeight;
 
+  /// Spacing between indicators.
   /// 指示器之间的间距
   final double indicatorMargin;
 
+  /// Distance of indicators from the bottom.
   /// 指示器距离底部的距离
   final double indicatorToBottom;
 
-  /// 默认指示器的形状，默认为矩形
+  /// Default indicator shape, defaults to a rectangle.
+  /// 默认指示器形状，默认为矩形
   final BoxShape indicatorShape;
 
-  /// 是否启用页面快速吸附，默认为 true
+  /// Whether to enable page snapping, defaults to true.
+  /// 是否启用页面快速吸附，默认为true
   final bool pageSnapping;
 
-  /// 是否在轮播首尾添加额外的页面， 默认为 true
+  /// Whether to add extra pages at the beginning and end of the carousel, default is true.
+  /// 是否在轮播首尾添加额外的页面，默认为true
   final bool padEnds;
 
+  /// The behavior for the clip.
+  /// 剪裁的行为
   final Clip clipBehavior;
 
+  /// Whether to reverse the order of items.
+  /// 是否倒序项目的顺序
   final bool reverse;
 
+  /// The axis along which the page view scrolls.
+  /// 页面视图滚动的轴
   final Axis scrollDirection;
 
+  /// Carousel controller, optional.
   /// 轮播控制器，可选
   final CarouselController? controller;
 
-  /// 是否禁用默认指示器回调函数
-  final bool disableIndicatorDefaultCallbacks = false;
+  /// Whether to disable default indicator callback functions, default is false.
+  /// 是否禁用默认指示器回调函数，默认为false。
+  final bool disableIndicatorDefaultCallbacks;
 
   const Carousel(
     this.units, {
@@ -111,6 +172,7 @@ class Carousel extends StatefulWidget {
     this.clipBehavior = Clip.hardEdge,
     this.reverse = false,
     this.scrollDirection = Axis.horizontal,
+    this.disableIndicatorDefaultCallbacks = false,
     this.controller,
   });