+++ lib/carousels.dart
@@ -1,4 +1,7 @@
 library awesome_carousel;
 
 export 'base/base.dart' show Carousel;
-export 'base/controller.dart';
+export 'base/controller.dart' show CarouselController;
+
+export 'img_carousels/net_img_carousel.dart' show NetImgCarousel;
+// import 'img_carousels/assert_img_carousel.dart';