+++ CHANGELOG.md
@@ -1,1 +1,15 @@
 # Changelog
+
+## [1.0.0+2] - 2023-10-17
+
+### Added
+
+Introduce network image carousel widget NetImgCarousel class.
+
+Introduce assets image carousel widget AssetCarousel class.
+
+## [1.0.0+1] - 2023-10-16
+
+### Added
+
+Introduce general carousel widget Carousel class.